Moved benchmark setup and teardown to models_test.go
